留言

+++ models/posts.js
@@ -1,5 +1,6 @@
 const marked = require("marked");
 const Post = require("../lib/mongo").Post;
+const CommentModel = require('./comments')
 
 // 将post的content从markdown转换成html
 Post.plugin("contentToHtml", {
@@ -17,6 +18,27 @@ Post.plugin("contentToHtml", {
   }
 });
 
+// 给post添加留言数commentsCount
+Post.plugin('addCommentsCount', {
+  afterFind: function(posts) {
+    return Promise.all(posts.map(function(post) {
+      return CommentModel.getCommentsCount(post._id).then(function(addCommentsCount) {
+        post.commentsCount = commentsCount
+        return post
+      })
+    }))
+  },
+  afterFindOne: function(post) {
+    if (post) {
+      return CommentModel.getCommentsCount(post._id).then(function(count) {
+        post.commentsCount = count
+        return post
+      })
+    }
+    return post
+  }
+})
+
 module.exports = {
   // 创建一篇文章
   create: function create(post) {
@@ -30,6 +52,7 @@ module.exports = {
         model: "User"
       })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec();
   },
@@ -43,6 +66,7 @@ module.exports = {
       .populate({ path: "author", model: "User" })
       .sort({ _id: -1 })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec();
   },
@@ -59,7 +83,12 @@ module.exports = {
     return Post.update({_id: postId}, {$set: data}).exec()
   },
   // 通过文章id删除一篇文章
-  delPostById: function delPostById(postId) {
-    return Post.deleteOne({_id: postId}).exec()
+  delPostById: function delPostById(postId, author) {
+    return Post.deleteOne({author: author, _id: postId}).exec().then(function(res) {
+      // 文章删除后，再删除该文章下的所有留言
+      if (res.result.ok && res.result.n > 0) {
+        return CommentModel.delCommentByPostId(postId)
+      }
+    })
   }
 };
